refactors dashboard creation
